requirements.txt
python-qpid-proton>=0.24.0
Flask>=0.11.0
requests